# src/main/resources/assets/api-docs.yaml
swagger: '2.0'
info:
  title: Consent
  description: A set of web-services containing methods for ingesting and retrieving expressions representing data use restrictions found within consent forms.
  version: 1.0.1
  license:
    name: BSD
    url: 'http://opensource.org/licenses/BSD-3-Clause'
  termsOfService: 'https://github.com/DataBiosphere/consent'
securityDefinitions:
  consent_auth:
    type: oauth2
    authorizationUrl: 'https://accounts.google.com/o/oauth2/auth'
    flow: implicit
    scopes:
      profile: profile
      email: email
paths:
...
  /api/consent/{consentId}/vote/{id}:
    post:
      summary: firstVoteUpdate
      description: Returns the created vote.
      parameters:
        - name: consentId
          in: path
          description: The consent ID.
          required: true
          type: string
        - name: id
          in: path
          description: The vote ID.
          required: true
          type: integer
          format: int32
        - name: rec
          in: body
          required: true
          schema:
            $ref: '#/definitions/Vote'
      tags:
        - Vote
      responses:
        200:
          description: Returns the created vote.
          schema:
            $ref: '#/definitions/Vote'
        400:
          description: 'Bad request, malformed entity, etc'
        404:
          description: The association type wasn't send, or the application couldn't find any consents that matched the search criteria.
    put:
      summary: updateConsentVote
      description: Returns the updated vote.
      parameters:
        - name: consentId
          in: path
          description: The consent ID.
          required: true
          type: string
        - name: id
          in: path
          description: The vote ID.
          required: true
          type: integer
          format: int32
        - name: rec
          in: body
          required: true
          schema:
            $ref: '#/definitions/Vote'
      tags:
        - Vote
      responses:
        200:
          description: Returns the created vote.
          schema:
            $ref: '#/definitions/Vote'
        400:
          description: Bad request, malformed entity, etc.
    get:
      summary: describe
      description: Returns the required vote.
      parameters:
        - name: consentId
          in: path
          description: The consent ID.
          required: true
          type: string
        - name: id
          in: path
          description: The vote ID.
          required: true
          type: integer
          format: int32
      tags:
        - Vote
      responses:
        200:
          description: Returns the required vote.
          schema:
            $ref: '#/definitions/Vote'
    delete:
      summary: deleteVote
      description: Deletes the required vote.
      parameters:
        - name: consentId
          in: path
          description: The consent ID.
          required: true
          type: string
        - name: id
          in: path
          description: The vote ID.
          required: true
          type: integer
          format: int32
      tags:
        - Vote
      responses:
        200:
          description: Response = Vote was deleted
          schema:
            type: string
        404:
          description: The vote couldn't be found
...
  /api/dataRequest/{requestId}/vote/{id}:
    get:
      summary: describe
      description: Returns the Vote with the specified Id.
      parameters:
        - name: requestId
          in: path
          description: referenceId of the Election of the requested Vote ( this value is not used in the search , its only to complete the path)
          required: true
          type: string
        - name: id
          in: path
          description: Vote ID
          required: true
          type: integer
          format: int32
      tags:
        - Vote
        - DataRequest
      responses:
        200:
          description: Returns the requested Vote.
          schema:
            $ref: '#/definitions/Vote'
        404:
          description: The Vote doesn't exist.
    post:
      summary: createDataRequestVote
      description: update a Vote specified by VoteId of a DataRequest Election.
      parameters:
        - name: rec
          in: body
          description: Vote to be created
          required: true
          schema:
            $ref: '#/definitions/Vote'
        - name: requestId
          in: path
          description: referenceId of the Election of the requested Vote ( this value is not used in the create process , its only to complete the path)
          required: true
          type: string
        - name: id
          in: path
          description: Vote ID
          required: true
          type: integer
          format: int32
      tags:
        - Vote
        - DataRequest
      responses:
        201:
          description: Returns the updated Vote.
        400:
          description: Required parameter is null
        404:
          description: The \"to update\" vote doesn't exist.
    put:
      summary: updateDataRequestVote
      description: update an edited Vote specified by VoteId of a DataRequest Election.
      parameters:
        - name: rec
          in: body
          description: Vote to be created
          required: true
          schema:
            $ref: '#/definitions/Vote'
        - name: requestId
          in: path
          description: referenceId of the Election of the requested Vote ( this value is not used in the create process , its only to complete the path)
          required: true
          type: string
        - name: id
          in: path
          description: Vote ID
          required: true
          type: integer
          format: int32
      tags:
        - Vote
        - DataRequest
      responses:
        201:
          description: Returns the updated Vote.
        400:
          description: Required parameter is null
        404:
          description: The \"to update\" vote doesn't exist.
    delete:
      summary: deleteVote
      description: Deletes the vote specified by the requestId and the VoteId.
      parameters:
        - name: requestId
          in: path
          description: referenceId of the Election at which the vote belongs.
          required: true
          type: string
        - name: id
          in: path
          description: Vote ID
          required: true
          type: integer
          format: int32
      tags:
        - Vote
        - DataRequest
      responses:
        200:
          description: The Vote was deleted.
        404:
          description: Could not find vote with this Vote ID.
  /api/dataRequest/{requestId}/vote/{id}/final:
    post:
      summary: updateFinalAccessConsentVote
      description: update a the FinalAccessVote specified by VoteId of a DataRequest Election and set finalAccessVote Election property to TRUE.
      parameters:
        - name: rec
          in: body
          description: Vote to be created
          required: true
          schema:
            $ref: '#/definitions/Vote'
        - name: requestId
          in: path
          description: referenceId of the Election of the requested Vote ( this value is not used in the process , its only to complete the path)
          required: true
          type: string
        - name: id
          in: path
          description: Vote ID
          required: true
          type: integer
          format: int32
      tags:
        - Vote
        - DataRequest
      responses:
        200:
          description: Returns the updated Vote.
          schema:
            $ref: '#/definitions/Vote'
        400:
          description: Required parameter is null
        404:
          description: Election for specified id does not exist
  /api/dataRequest/{requestId}/vote/final:
    get:
      summary: describeFinalAccessVote
      description: Returns the Vote with isfinalAccessVote property  = TRUE with the specified Id.
      parameters:
        - name: requestId
          in: path
          description: Vote ID
          required: true
          type: integer
          format: int32
      tags:
        - Vote
        - DataRequest
      responses:
        200:
          description: Returns the requested Vote.
          schema:
            $ref: '#/definitions/Vote'
        404:
          description: Could not find a vote with this id.
...
definitions:
...
  Vote:
    type: object
    properties:
      voteId:
        type: integer
        format: int32
        description: Describes the id of the vote.
      vote:
        type: boolean
        description: Describe the Positive or negative value of the vote
      dacUserId:
        type: integer
        format: int32
        description: Describes the id of the voter.
      createDate:
        type: string
        format: date
        description: Describes the date the vote was created.
      updateDate:
        type: string
        format: date
        description: Describes the date the vote was last modified.
      electionId:
        type: integer
        format: int32
        description: Describes the id of the election the election at which the vote belongs.
      rationale:
        type: string
        description: If the final vote is NO, the voter can specify why in this field.
      type:
        type: string
        description: Specifies the type of the vote (DAC, CHAIRPERSON, AGREEMENT, FINAL)
      isReminderSent:
        type: boolean
        description: Describes if the user recieved a reminder by the chairperson to log the vote on the case.
      hasConcerns:
        type: boolean
        description: For Dataset elections, it is a flag for DataSet owners concerns.
...